+++ src/single_turn/single_turn_trait.rs
@@ -1,0 +1,12 @@
+use async_trait::async_trait;
+use crate::errors::ModelError;
+
+#[async_trait]
+pub trait SingleTurnLlm {
+    fn name(&self) -> &str;
+    async fn generate(
+        &self,
+        prompt: String,
+        preprompt: Option<String>,
+    ) -> Result<String, ModelError>;
+}
